requirements.txt
asgiref==3.8.1
astroid==3.2.4
async-timeout==4.0.3
boto3==1.34.153
botocore==1.34.153
colorama==0.4.6
dill==0.3.8
dj-database-url==2.2.0
Django==5.0.7
django-filter==24.3
django-redis==5.4.0
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
isort==5.13.2
jmespath==1.0.1
mccabe==0.7.0
pillow==10.4.0
platformdirs==4.2.2
psycopg2==2.9.9
PyJWT==2.9.0
pylint==3.2.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-logstash==0.4.8
redis==5.0.8
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
tomlkit==0.13.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
